+++ src/components/msg/index.ts
@@ -62,6 +62,10 @@ import ConnectionOpenTry from './ibc/connection_open_try';
 import CounterpartyConnection from './ibc/connection_counterparty';
 import Version from './ibc/connection_version';
 import Transfer from './ibc_transfer/transfer';
+import Grant from './authz/grant';
+import Revoke from './authz/revoke';
+import GrantAllowance from './feegrant/grant_allowance';
+import RevokeAllowance from './feegrant/revoke_allowance';
 
 export {
   getMessageModelByType,
@@ -122,4 +126,8 @@ export {
   CounterpartyConnection,
   Version,
   Transfer,
+  Grant,
+  Revoke,
+  GrantAllowance,
+  RevokeAllowance,
 };